src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
black==25.9.0
boto3==1.40.69
botocore==1.40.69
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
copulas==0.12.3
coverage==7.11.1
ctgan==0.11.1
cycler==0.12.1
dataclasses-json==0.6.7
deepecho==0.7.0
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
Faker==37.12.0
fastavro==1.12.1
fastuuid==0.14.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
litellm==1.79.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.7
llama-index-cli==0.5.3
llama-index-core==0.14.7
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.7
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.11.1
llama-parse==0.6.54
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.10.2
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
openai==1.109.1
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
plotly==6.4.0
pluggy==1.6.0
ply==3.11
propcache==0.4.1
psycopg2-binary==2.9.11
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.3
pyparsing==3.2.5
pypdf==6.1.3
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
rdt==1.18.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
ruff==0.14.4
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.3
sdmetrics==0.24.0
sdv==1.28.0
shellingham==1.5.4
simple-ddl-parser==1.7.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
